--- LevelsMenu.cs
@@ -1,42 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
-
-namespace ReadingTree
-{
-    public partial class LevelsMenu : Form
-    {
-        public LevelsMenu()
-        {
-            InitializeComponent();
-        }
-
-        private void label4_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void LevelsMenu_Load(object sender, EventArgs e)
-        {
-
-        }
-
-        private void headerName_Enter(object sender, EventArgs e)
-        {
-
-        }
-
-        private void btnMainMenu_Click(object sender, EventArgs e)
-        {
-            MainMenu main = new MainMenu();
-            main.Show();
-            this.Hide();
-        }
-    }
-}
