--- PossessiveNounMenu.cs
@@ -1,25 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
-
-namespace ReadingTree
-{
-    public partial class PossessiveNounMenu : Form
-    {
-        public PossessiveNounMenu()
-        {
-            InitializeComponent();
-        }
-
-        private void PossessiveNounMenu_Load(object sender, EventArgs e)
-        {
-
-        }
-    }
-}
