+++ ReadingTree/DigraphMenu.cs
@@ -1,0 +1,32 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Data;
+using System.Drawing;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Forms;
+
+namespace ReadingTree
+{
+    public partial class DigraphMenu : Form
+    {
+        public DigraphMenu()
+        {
+            InitializeComponent();
+        }
+
+        private void Digraph_Load(object sender, EventArgs e)
+        {
+
+        }
+
+        private void btnMainMenu_Click(object sender, EventArgs e)
+        {
+            MainMenu main = new MainMenu();
+            main.Show();
+            Hide();
+        }
+    }
+}
