--- VCVariablesMenu.cs
@@ -1,35 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
-
-namespace ReadingTree
-{
-    public partial class VCVariablesMenu : Form
-    {
-        public VCVariablesMenu()
-        {
-            InitializeComponent();
-        }
-
-        private void VCVariablesMenu_Load(object sender, EventArgs e)
-        {
-
-        }
-
-        private void button8_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void VCVariablesLabel_Click(object sender, EventArgs e)
-        {
-
-        }
-    }
-}
